+++ clients/admin-ui/src/features/data-discovery-and-detection/tables/cells/DetectionItemActionsCell.tsx
@@ -62,7 +62,7 @@ const DetectionItemActionsCell = ({
     childDiffHasChanges;
 
   return (
-    <HStack onClick={(e) => e.stopPropagation()}>
+    <HStack>
       {showStartMonitoringAction && (
         <ActionButton
           title="Monitor"