+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Added support for AWS SES as an email provider [#5804](https://github.com/ethyca/fides/pull/5804)
 - Nested identity query support for BigQuery [#5814](https://github.com/ethyca/fides/pull/5814)
 - Added job that automatically requeues interrupted tasks for in progress privacy requests [#5800](https://github.com/ethyca/fides/pull/5800)
+- Added "Assets" tab on system view for web monitor assets [#5811](https://github.com/ethyca/fides/pull/5811)
 - DB model support for Attachments [#5784](https://github.com/ethyca/fides/pull/5784) https://github.com/ethyca/fides/labels/db-migration
 
 ### Changed
@@ -34,6 +35,9 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 
 ### Docs
 - Removed version pins in LDFLAGS & CFLAGS for local MSSQL builds [#5760](https://github.com/ethyca/fides/pull/5760)
+
+### Fixed
+- Fixed bug with D&D table column widths [#5813](https://github.com/ethyca/fides/pull/5813)
 
 ## [2.55.4](https://github.com/ethyca/fides/compare/2.55.3...2.55.4)